refactor ♻️ : migrate to pnpm workspace
