+++ .github/pull_request_template.md
@@ -1,0 +1,12 @@
+## What
+
+<!-- Brief description of the change -->
+
+## Linear
+
+Closes <!-- PER-XX -->
+
+## Checklist
+
+- [ ] `pnpm validate` passes
+- [ ] Tested locally
