+++ .gitignore
@@ -3,3 +3,4 @@ node_modules
 cms/.env
 !.env.sample
 .output
+.claude
