+++ cms/package.json
@@ -13,6 +13,7 @@
     "lint:fix": "biome check --write .",
     "format": "biome format --write .",
     "format:check": "biome format --check .",
+    "validate": "tsc --noEmit && biome check .",
     "deploy": "sanity deploy",
     "deploy-graphql": "sanity graphql deploy"
   },
@@ -29,27 +30,7 @@
   "devDependencies": {
     "@biomejs/biome": "2.4.16",
     "@types/react": "^19.1",
-    "typescript": "^6.0.0"
-  },
-  "pnpm": {
-    "onlyBuiltDependencies": [
-      "esbuild"
-    ],
-    "overrides": {
-      "@babel/plugin-transform-modules-systemjs": "7.29.7",
-      "ajv": "^8.0.0",
-      "flatted": "3.4.2",
-      "minimatch@10>brace-expansion": "5.0.6",
-      "tinyglobby>picomatch": "4.0.4",
-      "filelist>minimatch": "10.2.5",
-      "jsdom>undici": "8.3.0",
-      "readdir-glob>minimatch": "10.2.5",
-      "@actions/github>undici": "8.2.0",
-      "@actions/http-client>undici": "8.2.0",
-      "vite": "8.0.16",
-      "follow-redirects": "1.16.0",
-      "dompurify": "3.4.7",
-      "@vercel/frameworks>js-yaml": "4.1.1"
-    }
+    "typescript": "^6.0.0",
+    "zod": "^4.1.12"
   }
 }
